Implemented suggestions and cutscene code
I didn't end up using MinDuration as suggested, as the outer if statement there means it wouldn't work. All other suggestions are implemented, though.

I added a new SkipCutscene field to Tutorial_Movement_PlayCutScene.
If disabled, it will play a test cutscene for now, which can also be played by setting the CutSceneToPlayOnStart field of the CutScenePlayer object in the scene. Just drag the TestCutScene object into that field (its a child object of the CutScenePlayer).