Jelly Level Affects Number of Jellydew Produced
Modified the DewInstantiate.cs script to add more than one dew depending on the value passed to the amount variable. Added two arrays to the Parameter.cs file for spawning more dew at the specific level and for analyzing the experience required to reach the next level. Added exception handling for if the slimes level is outside of range of the two arrays.