Throw Item When Tossed
Items should now be thrown when tossed. I added some functionality where the block spawns above the player if they are in front of a block and tossing the item. I also added a timer to the PickupItem.cs that enables the isTrigger property variable when the item is thrown. This is only for items spawned into the game though.